lint: fix all clippy warnings and stalemate bitboard index bug
- Fix 64 clippy warnings: unused imports, deprecated rand 0.9 APIs
  (thread_rng → rng), unnecessary casts, manual_range_contains,
  needless_range_loop, dead code annotations
- Fix calculate_stalemate() using bitboard value as array index
  instead of square index (.trailing_zeros())

+++ src/game/playerobj.rs
@@ -51,11 +51,11 @@ impl Player {
             panic!("Invalid move notation: {}", notation);
         }
 
-        let from_file = (chars[0] as u8 - b'a') as u8;
+        let from_file = chars[0] as u8 - b'a';
         let from_rank = (chars[1].to_digit(10).unwrap() - 1) as u8;
         let from_idx = from_rank * 8 + from_file;
 
-        let to_file = (chars[2] as u8 - b'a') as u8;
+        let to_file = chars[2] as u8 - b'a';
         let to_rank = (chars[3].to_digit(10).unwrap() - 1) as u8;
         let to_idx = to_rank * 8 + to_file;
